Move loop/spin/rotate to new effect, from shape.
