+++ src/features/exporter/CommentExporter.js
@@ -43,34 +43,15 @@ export default class CommentExporter extends Exporter {
           this.optionLines(shape, layer, Object.keys(options))
           this.keyValueLine('Visible', layer.visible)
           this.optionLines(transform, layer, ['startingWidth', 'startingHeight', 'offsetX', 'offsetY', 'rotation', 'reverse'])
-          this.optionLines(transform, layer, ['numLoops', 'transformMethod'], layer.repeatEnabled)
 
-          this.optionLine(transform, layer, 'growEnabled', layer.repeatEnabled)
-          this.indent()
-          this.optionLine(transform, layer, 'growValue', layer.repeatEnabled && layer.growEnabled)
-          this.optionLine(transform, layer, 'growMethod', layer.repeatEnabled && layer.growEnabled)
-          this.indent()
-          this.optionLine(transform, layer, 'growMath', layer.repeatEnabled && layer.growEnabled && layer.growMethod === 'function')
-          this.dedent()
-          this.dedent()
-
-          this.optionLine(transform, layer, 'spinEnabled', layer.repeatEnabled)
-          this.indent()
-          this.optionLines(transform, layer, ['spinValue', 'spinMethod'], layer.repeatEnabled && layer.spinEnabled)
-          this.indent()
-          this.optionLine(transform, layer, 'spinMath', layer.repeatEnabled && layer.spinEnabled && layer.spinMethod === 'function')
-          this.optionLine(transform, layer, 'spinSwitchbacks', layer.repeatEnabled && layer.spinEnabled && layer.spinMethod === 'constant')
-          this.dedent()
-          this.dedent()
-
-          this.optionLine(transform, layer, 'trackEnabled', layer.repeatEnabled)
-          this.indent()
-          this.optionLines(transform, layer, ['trackValue', 'trackLength', 'trackNumLoops'], layer.repeatEnabled && layer.trackEnabled)
-          this.optionLine(transform, layer, 'trackGrowEnabled', layer.repeatEnabled && layer.trackEnabled)
-          this.indent()
-          this.optionLine(transform, layer, 'trackGrow', layer.repeatEnabled && layer.trackGrowEnabled)
-          this.dedent()
-          this.dedent()
+          // this.optionLine(transform, layer, 'trackEnabled', layer.repeatEnabled)
+          // this.indent()
+          // this.optionLines(transform, layer, ['trackValue', 'trackLength', 'trackNumLoops'], layer.repeatEnabled && layer.trackEnabled)
+          // this.optionLine(transform, layer, 'trackGrowEnabled', layer.repeatEnabled && layer.trackEnabled)
+          // this.indent()
+          // this.optionLine(transform, layer, 'trackGrow', layer.repeatEnabled && layer.trackGrowEnabled)
+          // this.dedent()
+          // this.dedent()
 
           if (!layer.effect) {
             this.line('Fine tuning:')
