+++ src/models/Shape.js
@@ -30,19 +30,7 @@ export default class Shape {
       offsetY: 0.0,
       open: true,
       rotation: 0,
-      numLoops: 10,
-      transformMethod: 'smear',
-      growEnabled: true,
-      growValue: 100,
-      growMethod: 'constant',
-      growMathInput: 'i+cos(i/2)',
-      growMath: 'i+cos(i/2)',
-      spinEnabled: false,
-      spinValue: 2,
-      spinMethod: 'constant',
-      spinMathInput: '10*sin(i/4)',
-      spinMath: '10*sin(i/4)',
-      spinSwitchbacks: 0,
+      numLoops: 1,
       trackEnabled: false,
       trackGrowEnabled: false,
       trackValue: 10,
